Created attributes, chenged every short to int

+++ SuperDUPER/Attack.cpp
@@ -11,7 +11,7 @@ Attack::Attack()
 
 void Attack::execute(GameLoop* gameLoop, const std::vector<float>& args) const {
 	if (args.size() != 2)
-		throw InvalidNumArgs{ 2, static_cast<short>(args.size()) };
+		throw InvalidNumArgs{ 2, static_cast<int>(args.size()) };
 	gameLoop->getPlayer()->attack(Position<>{static_cast<int>(args[0]), static_cast<int>(args[1])});
 }
