+++ SuperDUPER/Entity.cpp
@@ -1,7 +1,7 @@
 #include "Entity.h"
 
 
-Entity::Entity(float x, float y, short healthPoint, short mass, float bluntDamageMultiplier, float sharpDamageMultiplier)
+Entity::Entity(float x, float y, int healthPoint, int mass, float bluntDamageMultiplier, float sharpDamageMultiplier)
 	: healthPoint{ healthPoint }, mass { mass }, bluntDamageMultiplier{ bluntDamageMultiplier }, sharpDamageMultiplier{ sharpDamageMultiplier }
 {
 	position.x = x;
@@ -12,12 +12,12 @@ void Entity::render(SDL_Renderer* renderer, const Camera& camera) const {
 	shape->render(renderer, camera, Position<>{static_cast<long>(position.x), static_cast<long>(position.y)});
 }
 
-void Entity::takeBluntDamage(short amount) {
-	healthPoint -= static_cast<short>(amount*bluntDamageMultiplier);
+void Entity::takeBluntDamage(int amount) {
+	healthPoint -= static_cast<int>(amount*bluntDamageMultiplier);
 }
 
-void Entity::takeSharpDamage(short amount) {
-	healthPoint -= static_cast<short>(amount*sharpDamageMultiplier);
+void Entity::takeSharpDamage(int amount) {
+	healthPoint -= static_cast<int>(amount*sharpDamageMultiplier);
 }
 
 Entity::~Entity()
