+++ SuperDUPER/LifeForm.h
@@ -19,17 +19,17 @@ class LifeForm :
 {
 public:
 	//---Internal processing---
-	LifeForm(float x, float y, int speed, short healthPoint, short radius, short mass = 0, Friendliness friendliness = Friendliness::Neutral,
+	LifeForm(float x, float y, int speed, int healthPoint, int radius, int mass = 0, Friendliness friendliness = Friendliness::Neutral,
 		float facingDirection = 0, float rotatingSpeed = .1, int sightRange = 1000, float sightArea = 1);//The default constructor
 	void render(SDL_Renderer* renderer, const Camera& camera) const = 0;
 	bool refresh(const Map& map, const LifeFormList& lifeForms, float deltaTime) override;//Method to call each frame, return false if the player is still alive (return !isAlive())
 	//Return true if the destination is reached
 	void setRotatingDestination(const Destination& destination);
-	void checkCollision(const Map& map, Position<> position, short radius);
+	void checkCollision(const Map& map, Position<> position, int radius);
 
 	//---External information---
 	bool isInSight(const Position<float>& otherPos) const { return sight.isInSight(position, otherPos, facingDirection); }
-	short getDamageMultiplier() const { return strength; }
+	int getDamageMultiplier() const { return strength; }
 	int getHealthPoint() const { return healthPoint; }
 	int getSightRange() const { return sight.getSightRange(); }
 	float getSightArea() const { return sight.getSightArea(); }
@@ -54,7 +54,7 @@ private:
 	const int id;
 	static int idCount;//Each lifeForm has a unique ID
 	int actualSpeed;//Speed unit is : pixel/sec
-	short strength = 50;
+	int strength = 50;
 	Sight sight;
 	int baseSpeed;//Speed unit is : pixel/sec
 	float rotatingSpeed;
@@ -72,6 +72,6 @@ private:
 	bool renderVision = false;
 	bool renderHealth = true;
 protected:
-	short radius;
+	int radius;
 };
