+++ SuperDUPER/Move.cpp
@@ -12,7 +12,7 @@ Move::Move()
 
 void Move::execute(GameLoop* gameLoop, const std::vector<float>& args) const {
 	if (args.size() != 2 && args.size() != 0)
-		throw InvalidNumArgs{ 2, static_cast<short>(args.size()) };
+		throw InvalidNumArgs{ 2, static_cast<int>(args.size()) };
 	if (args.size() == 0)
 		gameLoop->getConsole()->addText(std::to_string(gameLoop->getPlayer()->getPosition().x) + " " + std::to_string(gameLoop->getPlayer()->getPosition().y));
 	else
